Improve year validation
- Create exception handler;
- Remove ErrorMessage schema;
- Create ValidationError exception;
- Refactor CustomPagination.
