+++ work_at_olist/base/books.py
@@ -1,6 +1,7 @@
 from django.http import HttpResponse
 from django.shortcuts import get_object_or_404
 
+from work_at_olist.base.exceptions import ValidationError
 from work_at_olist.base.models import Author, Book
 from work_at_olist.base.schemas import BookIn
 
@@ -10,7 +11,7 @@ def create_book(payload_dict, response: HttpResponse):
         get_object_or_404(Author, id=author_id)
 
     if not 1 < payload_dict['publication_year'] < 9999:
-        return 400, {'message': 'Please, fill a valid year.'}
+        raise ValidationError('Please, fill a valid publication year.')
 
     authors = payload_dict.pop('authors')
     book = Book.objects.create(**payload_dict)
