+++ work_at_olist/base/exceptions.py
@@ -1,0 +1,2 @@
+class ValidationError(Exception):
+    pass
